Extract user data purge into src/core/purge.ts
Move moderator/self data deletion, pending withdrawal, and verification
data removal into a dedicated module. Typecheck, 167 tests, and build
all pass.
